# src/dorc-web/src/apis/dorc-api/models/UserDbPermissionApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface UserDbPermissionApiModel
 */
export interface UserDbPermissionApiModel {
    /**
     * @type {number}
     * @memberof UserDbPermissionApiModel
     */
    UserId?: number;
    /**
     * @type {string}
     * @memberof UserDbPermissionApiModel
     */
    UserLoginId?: string | null;
    /**
     * @type {string}
     * @memberof UserDbPermissionApiModel
     */
    UserLoginType?: string | null;
    /**
     * @type {number}
     * @memberof UserDbPermissionApiModel
     */
    PermissionId?: number;
    /**
     * @type {string}
     * @memberof UserDbPermissionApiModel
     */
    PermissionName?: string | null;
    /**
     * @type {string}
     * @memberof UserDbPermissionApiModel
     */
    PermissionDisplayName?: string | null;
    /**
     * @type {number}
     * @memberof UserDbPermissionApiModel
     */
    DbId?: number;
    /**
     * @type {string}
     * @memberof UserDbPermissionApiModel
     */
    Tags?: string | null;
}