# src/dorc-web/src/apis/dorc-api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Dorc.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/RefDataDatabaseUsers": {
      "get": {
        "tags": [
          "RefDataDatabaseUsers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "envId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserApiModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserApiModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/RefDataDatabaseUsers/GetDbUsersPermissions": {
      "get": {
        "tags": [
          "RefDataDatabaseUsers"
        ],
        "parameters": [
          {
            "name": "serverName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "databaseName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDbPermissionApiModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDbPermissionApiModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserApiModel": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "LoginId": {
            "type": "string",
            "nullable": true
          },
          "DisplayName": {
            "type": "string",
            "nullable": true
          },
          "LoginType": {
            "type": "string",
            "nullable": true
          },
          "LanId": {
            "type": "string",
            "nullable": true
          },
          "Team": {
            "type": "string",
            "nullable": true
          },
          "LanIdType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserDbPermissionApiModel": {
        "type": "object",
        "properties": {
          "UserId": {
            "type": "integer",
            "format": "int32"
          },
          "UserLoginId": {
            "type": "string",
            "nullable": true
          },
          "UserLoginType": {
            "type": "string",
            "nullable": true
          },
          "PermissionId": {
            "type": "integer",
            "format": "int32"
          },
          "PermissionName": {
            "type": "string",
            "nullable": true
          },
          "PermissionDisplayName": {
            "type": "string",
            "nullable": true
          },
          "DbId": {
            "type": "integer",
            "format": "int32"
          },
          "Tags": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://identityserver:5200/connect/authorize",
            "tokenUrl": "https://identityserver:5200/connect/token",
            "scopes": {
              "dorc-api-np.manage": "Access to DORC API"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "dorc-api-np.manage"
      ]
    }
  ]
}